refactor(pdf): extract ledger validation
